Add pre-commit and helper script to update changelog
update-pkg-changelog.sh opens the changelog file of a package
with a new entry ready to be filled in. It is basically a helper
which executes dch with the proper flags (it depends on dch available
in the system). This is a standalone script, i.e. it can be used without the
pre-commit hook described bellow.

pre-commit.sh can be added as a git pre-commit hook to create a new
changelog entry whenever a package is modified by a new commit.