Add build-pkg.sh script
The goal of this script is to standardize how packages are generated and
patched, as well as allow an easy path to build a single package instead
of triggering the build of all packages everytime. Also, downloaded
source files and build directories are isolated from what is commited to
git, avoiding contamination and bringing backe the joy of a git clean
that effectively cleans up the local clone.

It worth mentioning that build-pkg.sh builds aero-* packages, the kernel
package will be handled by a separate build step in an upcoming commit.

This script has a configuration file at scripts/conf where repos' URLs and
hashes are defined. Those will be referenced by each package conf file
(e.g. aero-init/conf).

Changes required as part of the packaging process will transition from
the current sed commands in Makefile to patches added to the package
directory, in a directory called patches (e.g. aero-init/patches).