Add build-kernel script to move kernel build steps out of Makefile
Initial version of the build-kernel script. It basically does what
the old Makefile did, including questionable steps like yes to all
questions made by make oldconfig (It has to change).

As the Linux kernel version won't change anytime soon for Aero, versions
are currently hardcoded in the script. This is something that can be
improved to make it futute proof.

The kernel build also happens in a build dir and the generated packages
are placed with the other aero-* deb filed.